+++ fixtures/Name.php
@@ -25,8 +25,8 @@ final class Name
     public static function strings(): Set
     {
         return Set\Strings::madeOf(
-            Set\Integers::between(1, 46)->map(\chr(...)),
-            Set\Integers::between(48, 127)->map(\chr(...)),
+            Set\Integers::between(32, 46)->map(\chr(...)),
+            Set\Integers::between(48, 126)->map(\chr(...)),
         )
             ->between(1, 255)
             ->filter(