+++ fixtures/Name.php
@@ -27,8 +27,10 @@ final class Name
         return Set\Strings::madeOf(
             Set\Integers::between(32, 46)->map(\chr(...)),
             Set\Integers::between(48, 126)->map(\chr(...)),
+            Set\Unicode::emoticons(),
         )
             ->between(1, 255)
+            ->filter(static fn($name) => \mb_strlen($name, 'ASCII') <= 255)
             ->filter(
                 static fn(string $name): bool => $name !== '.' &&
                     $name !== '..' &&