+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Special ASCII characters are no longer used in genarated names via `Fixtures\Innmind\Filesystem\Name::any()`
+
 ## 7.4.0 - 2023-12-02
 
 ### Added